+++ README.md
@@ -1,2 +1,6 @@
 # text-summarization-web-app-in-flask
 A flask web app for text summarization. The app is based on these algorithms K-means, Lex Rank and Lsa which usage a NLP tokenizer implemented in python. A user can use any type of technique to get different types of result. the summerized text will be auto saved in a text file.
+
+#Before Running app.py
+
+sudo pip install sumy
